rename device_data to consts
